+++ extensions/remove-paywall/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Remove Paywall Changelog
 
-## [Update] - 2023-07-17
+## [Specify paywall remover] - 2023-07-24
 
-- Add archive.is
+- New command: "Remove Paywall Using"
+- Support for URLs in the clipboard
+
+## [New service] - 2023-07-17
+
+- Added Archive as a paywall removal service
 
 ## [Initial Version] - 2023-06-03